Multiplayer, player and title screen
In this commit, i have installed packages for foundation on the game. includes multiplayer modules, multiplayer testing modules. i have also added a character for temporarily, until final character is made. i have aslo started making the title screen for the  game, it is on hold until we decide a title.