changes to multiplyer, title, buttons, loading into the game, etc.
In this commit, I have done. completed the title page of our game. It's not complete yet, but the. process of loading in into the game is being completed. I am. easily able to join in into the main world from the title screen and I have added a picture for randomness. Dark souls. I have tweet the multiplayer system and player UI but it still requires testing and some improvements It is working but not as effectively. I have created new scenes. which complement into the transition from title screen to the main game