Input manager, centralization concept,
In this commit i have completed the input system and i have only completed mapping the movement (WASD). i have tested it works.
for centralization, my theory is the script structure that will be on our characters in the future. We're going to have components like the character animate manager, the character of locomotion manager, the character combat manager,etc.  Now we don't want to have to call a copy of each scripts on each of these components separately like if you want to reference the character animator from the combat manager, you don't want to have to write the character animator variable on the combat manager and then write it again in the locomotion manager and then write it again on another scripts we will make. So my idea is to centralize all of these actions in one place. That can be done through character player manager. It's basically going to hold every component that the character needs. And if another component needs to reference any of these, they get it through this script.