Player, Entities
Adds a simple interface component for all entity objects, including (Speed, Friction, Health) as some methods for interacting with it.

Adds the player prefab which uses the entity controller component and a player controller component which handles basic movement and dash mechanics.

(The player does not use a state machine but we may need to refactor it later to use one.)